Re-organises disk module
